+++ examples/quickstart.rb
@@ -16,16 +16,16 @@ actions = {
   :say => -> (session_id, msg) {
     p msg
   },
-  :merge => -> (context, entities) {
+  :merge => -> (session_id, context, entities, msg) {
     new_context = context.clone
     loc = first_entity_value entities, 'location'
     new_context['loc'] = loc unless loc.nil?
     return new_context
   },
-  :error => -> (session_id, msg) {
+  :error => -> (session_id, context) {
     p 'Oops I don\'t know what to do.'
   },
-  :'fetch-weather' => -> (context) {
+  :'fetch-weather' => -> (session_id, context) {
     new_context = context.clone
     new_context['forecast'] = 'sunny'
     return new_context