+++ lib/wit.rb
@@ -37,9 +37,9 @@ def validate_actions(actions)
     raise WitException.new "The '#{k}' action name should be a symbol" unless k.is_a? Symbol
     raise WitException.new "The '#{k}' action should be a lambda function" unless v.respond_to? :call and v.lambda?
     raise WitException.new "The \'say\' action should take 2 arguments: session_id, msg. #{learn_more}" if k == :say and v.arity != 2
-    raise WitException.new "The \'merge\' action should take 2 arguments: context, entities. #{learn_more}" if k == :merge and v.arity != 2
-    raise WitException.new "The \'error\' action should take 2 arguments: session_id, msg. #{learn_more}" if k == :error and v.arity != 2
-    raise WitException.new "The '#{k}' action should take 1 argument: context. #{learn_more}" if k != :say and k != :merge and k != :error and v.arity != 1
+    raise WitException.new "The \'merge\' action should take 4 arguments: session_id, context, entities, msg. #{learn_more}" if k == :merge and v.arity != 4
+    raise WitException.new "The \'error\' action should take 2 arguments: session_id, context. #{learn_more}" if k == :error and v.arity != 2
+    raise WitException.new "The '#{k}' action should take 2 arguments: session_id, context. #{learn_more}" if k != :say and k != :merge and k != :error and v.arity != 2
   end
   return actions
 end
@@ -78,7 +78,7 @@ class Wit
     req @access_token, Net::HTTP::Post, '/converse', params, context
   end
 
-  def run_actions(session_id, message, context={}, max_steps=DEFAULT_MAX_STEPS)
+  def run_actions_(session_id, message, context, max_steps, user_message)
     raise WitException.new 'max iterations reached' unless max_steps > 0
 
     rst = converse session_id, message, context
@@ -95,7 +95,7 @@ class Wit
     elsif type == 'merge'
       raise WitException.new 'unknown action: merge' unless @actions.has_key? :merge
       logger.info 'Executing merge'
-      context = @actions[:merge].call context, rst['entities']
+      context = @actions[:merge].call session_id, context, rst['entities'], user_message
       if context.nil?
         logger.warn 'missing context - did you forget to return it?'
         context = {}
@@ -104,7 +104,7 @@ class Wit
       action = rst['action'].to_sym
       raise WitException.new "unknown action: #{action}" unless @actions.has_key? action
       logger.info "Executing action #{action}"
-      context = @actions[action].call context
+      context = @actions[action].call session_id, context
       if context.nil?
         logger.warn 'missing context - did you forget to return it?'
         context = {}
@@ -112,10 +112,16 @@ class Wit
     elsif type == 'error'
       raise WitException.new 'unknown action: error' unless @actions.has_key? :error
       logger.info 'Executing error'
-      @actions[:error].call session_id, 'unknown action: error'
+      @actions[:error].call session_id, context
     else
       raise WitException.new "unknown type: #{type}"
     end
     return run_actions session_id, nil, context, max_steps - 1
   end
+
+  def run_actions(session_id, message, context={}, max_steps=DEFAULT_MAX_STEPS)
+    return run_actions_ session_id, message, context, max_steps, message
+  end
+
+  private :run_actions_
 end